requirements.txt
google-genai
cohere
pillow
numpy
streamlit